Remove unnecessary request spec stuff
