+++ spec/rails_helper.rb
@@ -29,9 +29,6 @@ RSpec.configure do |config|
   require 'support/active_admin_integration_spec_helper'
   config.include ActiveAdminIntegrationSpecHelper
 
-  require 'support/active_admin_request_helpers'
-  config.include ActiveAdminRequestHelpers, type: :request
-
   # Setup Some Admin stuff for us to play with
   config.before(:suite) do
     ActiveAdminIntegrationSpecHelper.load_defaults!
