+++ spec/requests/default_namespace_spec.rb
@@ -76,4 +76,16 @@ RSpec.describe ActiveAdmin::Application, type: :request do
 
     with_temp_application(application) { yield }
   end
+
+  def with_temp_application(application)
+    original_application = ActiveAdmin.application
+    ActiveAdmin.application = application
+    load_defaults!
+    reload_routes!
+
+    yield
+
+  ensure
+    ActiveAdmin.application = original_application
+  end
 end
